Dedicated Reservoir_Exponential_Outlet header.
Moving Reservoir_Exponential_Outlet class from Nonlinear_Reservoir.hpp
to dedicated header file.
